Updated: Changed totalItems from Int to Long

+++ src/main/kotlin/com/balladesh/tagggerapp/database/page/PagedResponse.kt
@@ -3,15 +3,14 @@ package com.balladesh.tagggerapp.database.page
 class PagedResponse<T>(
   val page: Int,
   val pageSize: Int,
-  private val totalItems: Int,
+  private val totalItems: Long,
   val itemsList: List<T>
 ) {
 
   val totalPages: Int
     get() {
-      val total: Int = if (pageSize == 0) { 1 } else { totalItems / pageSize }
+      val total: Int = if (pageSize == 0) { 1 } else { (totalItems / pageSize).toInt() }
 
       return if (total == 0) { 1 } else { total }
     }
-
 }
